refactor(cli): create seperate file for removePrePublishScript function
